Default starter weekday
* Minor bug fixes
