--- CVCalendar Demo/CVCalendar Demo/CVCalendar DemoTests/CVCalendar_DemoTests.swift
@@ -1,36 +1,0 @@
-//
-//  CVCalendar_DemoTests.swift
-//  CVCalendar DemoTests
-//
-//  Created by Мак-ПК on 1/3/15.
-//  Copyright (c) 2015 GameApp. All rights reserved.
-//
-
-import UIKit
-import XCTest
-
-class CVCalendar_DemoTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        XCTAssert(true, "Pass")
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock() {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
-}
