--- CVCalendar Demo/CVCalendar Demo/CVCalendar/CVCalendarViewDelegate.swift
@@ -1,19 +1,0 @@
-//
-//  CVCalendarViewDelegate.swift
-//  CVCalendar
-//
-//  Created by E. Mozharovsky on 12/27/14.
-//  Copyright (c) 2014 GameApp. All rights reserved.
-//
-
-import UIKit
-
-@objc
-protocol CVCalendarViewDelegate {
-    func shouldShowWeekdaysOut() -> Bool
-    func didSelectDayView(dayView: CVCalendarDayView)
-    func presentedDateUpdated(date: CVDate)
-
-    func dotMarker(shouldShowOnDayView dayView: CVCalendarDayView) -> Bool
-    func dotMarker(colorOnDayView dayView: CVCalendarDayView) -> UIColor
-}
