+++ CVCalendar Demo/CVCalendar Demo/Images.xcassets/present-100.imageset/Contents.json
