+++ CVCalendar Demo/CVCalendar Demo/Images.xcassets/present-100.imageset/present-100.png
