+++ CVCalendar Demo/CVCalendar Demo/Info.plist
@@ -36,7 +36,5 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
-	<key>CVCalendarStarterWeekday</key>
-	<integer>2</integer>
 </dict>
 </plist>
