+++ CVCalendar Demo/CVCalendar/CVCalendarManager.swift
@@ -25,14 +25,24 @@ class CVCalendarManager: NSObject {
     
     // MARK: - Private initialization
     
-    var starterWeekday = 2
+    var starterWeekday: Int?
     
     private override init() {
         self.calendar = NSCalendar(identifier: NSCalendarIdentifierGregorian)
         self.currentDate = NSDate()
         self.components = self.calendar?.components(NSCalendarUnit.MonthCalendarUnit | NSCalendarUnit.DayCalendarUnit, fromDate: self.currentDate!)
         
-        self.calendar?.firstWeekday = starterWeekday
+        let propertyName = "CVCalendarStarterWeekday"
+        let firstWeekday = NSBundle.mainBundle().objectForInfoDictionaryKey(propertyName) as? Int
+        if firstWeekday != nil {
+            self.starterWeekday = firstWeekday
+            self.calendar!.firstWeekday = starterWeekday!
+        } else {
+            let currentCalendar = NSCalendar.currentCalendar()
+            let firstWeekday = currentCalendar.firstWeekday
+            self.starterWeekday = firstWeekday
+            self.calendar!.firstWeekday = starterWeekday!
+        }
     }
     
     // MARK: - Common date analysis
